Addressing comments
- Remove types from Users.ts
- Move getParticipantsAwaitingApproval to service
