+++ src/api/services/participantsService.ts
@@ -1,7 +1,12 @@
 import { NextFunction, Request, Response } from 'express';
 import { z } from 'zod';
 
-import { Participant, ParticipantCreationPartial } from '../entities/Participant';
+import { Approver } from '../entities/Approver';
+import {
+  Participant,
+  ParticipantCreationPartial,
+  ParticipantStatus,
+} from '../entities/Participant';
 import { ParticipantType } from '../entities/ParticipantType';
 import { SSP_WEB_BASE_URL } from '../envars';
 import { getSharingList, SharingListResponse, updateSharingList } from './adminServiceClient';
@@ -51,6 +56,21 @@ export const fetchSharingParticipants = async (
 export const getSharingParticipants = async (participantSiteId: number): Promise<Participant[]> => {
   const sharingListResponse = await getSharingList(participantSiteId);
   return fetchSharingParticipants(sharingListResponse);
+};
+
+export const getParticipantsAwaitingApproval = async (email: string): Promise<Participant[]> => {
+  const approvers = await Approver.query().distinct('participantTypeId').where('email', email);
+  const approvableParticipantTypeIds = approvers.map((approver) => approver.participantTypeId);
+  const participantsAwaitingApproval = await Participant.query()
+    .whereIn(
+      'id',
+      Participant.relatedQuery('types')
+        .whereIn('participantTypeId', approvableParticipantTypeIds)
+        .select('participantId')
+    )
+    .withGraphFetched('types')
+    .where('status', ParticipantStatus.AwaitingApproval);
+  return participantsAwaitingApproval;
 };
 
 export const addSharingParticipants = async (
